pyproject.toml
[tool.poetry]
name = "simplecounter"
version = "0.1.0"
description = ""
authors = ["Rizal Zulfikar Rinanda <rzrinanda@github.com>"]

[tool.poetry.dependencies]
python = "3.10.1"
django-crispy-forms = "^1.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
